+++ AmplifyClients/AmplifyConnectClient/Tests/ConnectClientHostApp/ConnectClientHostApp/ContentView.swift
@@ -1,0 +1,14 @@
+//
+// Copyright Amazon.com Inc. or its affiliates.
+// All Rights Reserved.
+//
+// SPDX-License-Identifier: Apache-2.0
+//
+
+import SwiftUI
+
+struct ContentView: View {
+    var body: some View {
+        Text("Connect Client Integration Test Host")
+    }
+}